virt-config: Rename `deprecation` package to `featuregate`
Currently, the `deprecation` package is limited to manage FGs that are
deprecated (due to the feature being graduated or removed).

Rename the package to `featuregate` in order to generalize it and add to
it support for the active FGs (i.e. the ones representing features in
the Alpha and Beta stages).

Signed-off-by: Edward Haas <edwardh@redhat.com>